fixed install dependencies

+++ ts_scripts/install_dependencies.py
@@ -92,16 +92,9 @@ class Linux(Common):
         if os.system("wget --version") != 0 or args.force:
             os.system(f"{self.sudo_cmd}apt-get install -y wget")
 
-    def install_libgit2(self):
-        os.system(
-            f"wget https://github.com/libgit2/libgit2/archive/refs/tags/v1.3.0.tar.gz -O libgit2-1.3.0.tar.gz"
-        )
-        os.system(f"tar xzf libgit2-1.3.0.tar.gz")
-        os.system(f"cd libgit2-1.3.0 && cmake . && make && sudo make install && cd ..")
-        os.system(f"rm -rf libgit2-1.3.0 && rm libgit2-1.3.0.tar.gz")
-
     def install_maven(self):
-        os.system(f"{self.sudo_cmd}apt-get install -y maven")
+        if os.system("mvn --version") != 0 or args.force:
+            os.system(f"{self.sudo_cmd}apt-get install -y maven")
 
 
 class Windows(Common):
@@ -131,12 +124,14 @@ class Darwin(Common):
             os.system("brew install openjdk@17")
 
     def install_nodejs(self):
-        os.system("brew unlink node")
-        os.system("brew install node@14")
-        os.system("brew link --overwrite node@14")
+        if os.system("npm --v") != 0 or args.force:
+            os.system("brew unlink node")
+            os.system("brew install node@14")
+            os.system("brew link --overwrite node@14")
 
     def install_node_packages(self):
-        os.system(f"{self.sudo_cmd} ./ts_scripts/mac_npm_deps")
+        if os.system("npm --v") != 0 or args.force:
+            os.system(f"{self.sudo_cmd} ./ts_scripts/mac_npm_deps")
 
     def install_wget(self):
         if os.system("wget --version") != 0 or args.force:
